+++ lib/eve_online/esi/alliances.rb
@@ -9,6 +9,10 @@ module EveOnline
         response
       end
 
+      def etag
+        current_etag
+      end
+
       def scope; end
 
       def url