+++ spec/eve_online/esi/universe_ancestries_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-describe EveOnline::ESI::Ancestries do
+describe EveOnline::ESI::UniverseAncestries do
   specify { expect(subject).to be_a(EveOnline::ESI::Base) }
 
   specify { expect(described_class::API_ENDPOINT).to eq('https://esi.evetech.net/v1/universe/ancestries/?datasource=%<datasource>s') }
@@ -62,6 +62,14 @@ describe EveOnline::ESI::Ancestries do
 
       specify { expect { subject.ancestries }.to change { subject.instance_variable_get(:@ancestries) }.from(nil).to([ancestry]) }
     end
+  end
+
+  describe '#etag' do
+    let(:current_etag) { double }
+
+    before { expect(subject).to receive(:current_etag).and_return(current_etag) }
+
+    specify { expect(subject.etag).to eq(current_etag) }
   end
 
   describe '#scope' do