+++ spec/features/alliances_spec.rb
@@ -3,15 +3,33 @@
 require 'spec_helper'
 
 describe 'List all alliances' do
-  before { VCR.insert_cassette 'esi/alliances' }
+  context 'when etag not set' do
+    before { VCR.insert_cassette 'esi/alliances/alliances' }
 
-  after { VCR.eject_cassette }
+    after { VCR.eject_cassette }
 
-  subject { EveOnline::ESI::Alliances.new }
+    subject { EveOnline::ESI::Alliances.new }
 
-  specify { expect(subject.scope).to eq(nil) }
+    specify { expect(subject.not_modified?).to eq(false) }
 
-  specify { expect(subject.alliance_ids.size).to eq(3024) }
+    specify { expect(subject.scope).to eq(nil) }
 
-  specify { expect(subject.alliance_ids.first).to eq(1_354_830_081) }
+    specify { expect(subject.alliance_ids.size).to eq(3028) }
+
+    specify { expect(subject.alliance_ids.first).to eq(1_354_830_081) }
+
+    specify { expect(subject.etag).to eq('97f0c48679f2b200043cdbc3406291fc945bcd652ddc7fc11ccdc37a') }
+  end
+
+  context 'when etag is set' do
+    before { VCR.insert_cassette 'esi/alliances/alliances_with_etag' }
+
+    after { VCR.eject_cassette }
+
+    subject { EveOnline::ESI::Alliances.new(etag: '97f0c48679f2b200043cdbc3406291fc945bcd652ddc7fc11ccdc37a') }
+
+    specify { expect(subject.not_modified?).to eq(true) }
+
+    specify { expect(subject.etag).to eq('97f0c48679f2b200043cdbc3406291fc945bcd652ddc7fc11ccdc37a') }
+  end
 end