+++ spec/features/corporation_spec.rb
@@ -1,0 +1,53 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+describe 'Get corporation information' do
+  context 'when etag not set' do
+    let(:options) { { corporation_id: 98_468_592 } }
+
+    before { VCR.insert_cassette 'esi/corporations/98468592' }
+
+    after { VCR.eject_cassette }
+
+    subject { EveOnline::ESI::Corporation.new(options) }
+
+    specify { expect(subject.not_modified?).to eq(false) }
+
+    specify { expect(subject.scope).to eq(nil) }
+
+    specify do
+      expect(subject.as_json).to eq(alliance_id: 99_007_916,
+                                    ceo_id: 1_721_864_142,
+                                    corporation_url: 'http://',
+                                    creator_id: 1_721_864_142,
+                                    date_founded: '2016-07-11 14:22:17.000000000 +0000',
+                                    description: '',
+                                    etag: '9f1b3761e93aba36ceaea6328e62031ec777317c70b7439b665d93fe',
+                                    faction_id: nil,
+                                    home_station_id: 60_011_893,
+                                    member_count: 155,
+                                    name: 'Bullshit Bingo Club',
+                                    shares: 1000,
+                                    tax_rate: 0.1,
+                                    ticker: 'BUBIC')
+    end
+  end
+
+  context 'when etag is set' do
+    let(:options) do
+      {
+        corporation_id: 98_468_592,
+        etag: '9f1b3761e93aba36ceaea6328e62031ec777317c70b7439b665d93fe'
+      }
+    end
+
+    before { VCR.insert_cassette 'esi/corporations/98468592_with_etag' }
+
+    after { VCR.eject_cassette }
+
+    subject { EveOnline::ESI::Corporation.new(options) }
+
+    specify { expect(subject.not_modified?).to eq(true) }
+  end
+end